# Reduce memory usage in streams