This change avoids allocating a 16KB buffer for every
open stream.  Instead, the open streams share a pool of buffers,
allocating only as many as are needed, i.e. the number of
simultaneously active transfers.  Most streams are expected
to be idle most of the time, so this should reduce peak memory
requirements on heavily loaded servers and memory-constrained
clients.

This expands on the use of sync.Pool that is already in place
for UDP in client.go.

Additionally, this change removes all per-transfer allocations
from streams, and removes one allocation from the client's UDP
read and write methods.

The changes can be seen in benchmark results.
Master:
```
BenchmarkTCPThroughput-4               80113     13883 ns/op       576 mbps      80 B/op       2 allocs/op
BenchmarkTCPMultiplexing-4              3184    315071 ns/op                 150874 B/op    1080 allocs/op
BenchmarkShadowsocksClient_UDPWrite-4   3356    394884 ns/op                   1333 B/op      22 allocs/op
```

This branch:
```
BenchmarkTCPThroughput-4               77156     14325 ns/op       558 mbps       0 B/op       0 allocs/op
BenchmarkTCPMultiplexing-4              3735    310730 ns/op                  76862 B/op    1061 allocs/op
BenchmarkShadowsocksClient_UDPWrite-4  10000    343833 ns/op                   1297 B/op      21 allocs/op
```

Each op in BenchmarkTCPThroughput represents a 1 KB read and write
(client+server), all on one socket.  Each op in
BenchmarkTCPMultiplexing represents a 1 KB read and write on a new
socket.