@bemasc 2020-04-08
> One thing I don't get is why we need to change so many things.

I agree; this change is too big.  I'll break it up for review.

> The reader and writer already hold byte[] objects. Shouldn't we just change the `make()` calls for them (and the wiring for it)?

That would reduce the number of allocations (reusing old buffers for new sockets), but it would not reduce the amount of memory "in use" (currently 16KB x 2 for every open socket).

> I will need more time to review the changes to stream.go, as that's quite intrusive, and there were a bunch of corner cases and nuances I had issues with when adapting and changing the code.

Yeah.  I just realized that this PR has at least one corner case bug: if an app tries to send a single byte and wait for a response, the socket would stall waiting for more data before sending.