@fortuna 2020-04-08
How can we avoid 2x16KB per connection? We need one buffer for the reader and one for the writer.

About the bug, I ran into so many corner cases writing this code, it was kind of a pain. That's why I'm a bit nervous about changing too much. But I'd like to understand the reasoning behind only one 16KB buffer per connection.