@fortuna 2020-04-08
Oh, I see. You're returning the buffer to the pool after every block is read or written. I don't think we need that level of optimization, it will make the code too complex. Also, the read optimization will only work if the read aligns with the buffer size (otherwise you have a leftover).

Replacing the make calls with the buffer pool will probably be good enough. Unless we have evidence that the extra optimization will have significant impact, I'd rather avoid it (how many open connections are there anyway?)