@bemasc 2020-04-09
> Oh, I see. You're returning the buffer to the pool after every block is read or written. I don't think we need that level of optimization, it will make the code too complex. Also, the read optimization will only work if the read aligns with the buffer size (otherwise you have a leftover).

To be clear, it still "works" once the pending data has been written and `sr.leftover` is drained.  In other words, it only holds the buffer while there is backpressure between the two sockets.

> Replacing the make calls with the buffer pool will probably be good enough. Unless we have evidence that the extra optimization will have significant impact, I'd rather avoid it

OK, I'll try to figure out how to measure that.  

> (how many open connections are there anyway?)

I just did a quick test using chrome://net-export, and browsing a couple of heavy websites resulted in 256 open sockets (Chrome's socket pool size limit).  Browsers only close sockets when the socket pool is full (or when a socket has been idle for a long time), in order to reuse sockets for subsequent requests to the same domain.  That means that each user can easily have hundreds of sockets open through the proxy, most of which are idle.